+++ Sources/Controllers/ALKNewChatViewController.swift
@@ -131,7 +131,7 @@ extension ALKNewChatViewController: UITableViewDelegate, UITableViewDataSource {
             
             let storyboard = UIStoryboard.name(storyboard: UIStoryboard.Storyboard.createGroupChat, bundle: Bundle.applozic)
             if let vc = storyboard.instantiateViewController(withIdentifier: "ALKCreateGroupViewController") as? ALKCreateGroupViewController {
-                vc.setCurrentGroupSelected(groupName: "", groupProfileImg: nil, groupSelected: [ALKFriendViewModel](), delegate: self)
+                vc.setCurrentGroupSelected(groupId: NSNumber(value: 0), groupProfile: nil, delegate: self)
                 vc.addContactMode = .newChat
                 vc.configuration = configuration
                 navigationController?.pushViewController(vc, animated: true)